[VCS6] Reorganize ViewHandlers.
They should've been under VersionControl.Views. Also, pack all the default ones into one file.
