+++ main/src/addins/VersionControl/MonoDevelop.VersionControl/MonoDevelop.VersionControl.Views/DefaultVersionControlViewHandlers.cs
@@ -1,0 +1,88 @@
+﻿//
+// DefaultVersionControlViewHandlers.cs
+//
+// Author:
+//       Marius Ungureanu <marius.ungureanu@xamarin.com>
+//
+// Copyright (c) 2015 Xamarin, Inc (http://www.xamarin.com)
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+using MonoDevelop.Ide;
+using MonoDevelop.Ide.Gui;
+using MonoDevelop.Projects.Text;
+using MonoDevelop.VersionControl;
+
+namespace MonoDevelop.VersionControl.Views
+{
+	public class DefaultBlameViewHandler : IBlameViewHandler
+	{
+		public bool CanHandle (VersionControlItem item, DocumentView primaryView)
+		{
+			return (primaryView == null || primaryView.GetContent <ITextFile> () != null)
+				&& DesktopService.GetFileIsText (item.Path);
+		}
+
+		public IBlameView CreateView (VersionControlDocumentInfo info)
+		{
+			return new BlameView (info);
+		}
+	}
+
+	public class DefaultDiffViewHandler : IDiffViewHandler
+	{
+		public bool CanHandle (VersionControlItem item, DocumentView primaryView)
+		{
+			return (primaryView == null || primaryView.GetContent <ITextFile> () != null)
+				&& DesktopService.GetFileIsText (item.Path);
+		}
+
+		public IDiffView CreateView (VersionControlDocumentInfo info)
+		{
+			return new DiffView (info);
+		}
+	}
+
+	public class DefaultLogViewHandler : ILogViewHandler
+	{
+		public bool CanHandle (VersionControlItem item, DocumentView primaryView)
+		{
+			return true;
+		}
+
+		public ILogView CreateView (VersionControlDocumentInfo info)
+		{
+			return new LogView (info);
+		}
+	}
+
+	public class DefaultMergeViewHandler : IMergeViewHandler
+	{
+		public bool CanHandle (VersionControlItem item, DocumentView primaryView)
+		{
+			return (primaryView == null || primaryView.GetContent <ITextFile> () != null)
+				&& DesktopService.GetFileIsText (item.Path);
+		}
+
+		public IMergeView CreateView (VersionControlDocumentInfo info)
+		{
+			return new MergeView (info);
+		}
+	}
+}
+
